Cleanup source code

+++ loop_subscreen.c
@@ -277,9 +277,9 @@ void loop_subscreen_top_drawtime()
 {
     UserInterface_ClearTextBuffer_Line(26);
     char temp_time_string[32] = "          ";
-    uint16_t temp_h = (uint16_t)((system_frames_elapsed / FPS) / (3600l));
-    uint16_t temp_m = (uint16_t)(((system_frames_elapsed / FPS) % (3600l)) / 60);
-    uint16_t temp_s = (uint16_t)((system_frames_elapsed / FPS) % 60);
+    uint16_t temp_h = (system_frames_elapsed / FPS) / (3600l);
+    uint16_t temp_m = ((system_frames_elapsed / FPS) % 3600l) / 60;
+    uint16_t temp_s = (system_frames_elapsed / FPS) % 60;
 
     if ((system_frames_elapsed % (60 / V_MUL)) >= (FPS / 2))
     {
